CrystallographyBase

Documentation for CrystallographyBase.

CrystallographyBase.ReciprocalPoint

CrystallographyBase.cellvolume

CrystallographyBase.cellvolume

CrystallographyBase.cellvolume

CrystallographyBase.supercell

CrystallographyBase.supercell

CrystallographyBase.ReciprocalPoint — Type

ReciprocalPoint(x, y, z, w)

Represent a special point of the 3D Brillouin zone. Each of them has a weight w.

CrystallographyBase.cellvolume — Method

cellvolume(a, b, c, α, β, γ)

Calculates the cell volume from 6 cell parameters.

CrystallographyBase.cellvolume — Method

cellvolume(g::MetricTensor)

Calculates the cell volume from a MetricTensor.

CrystallographyBase.cellvolume — Method

cellvolume(l::Lattice)
cellvolume(c::Cell)

Calculates the cell volume from a Lattice or a Cell.

CrystallographyBase.supercell — Method

supercell(cell::Lattice, expansion::AbstractMatrix{<:Integer})

Allow the supercell to be a tilted extension of cell.

CrystallographyBase.supercell — Method

supercell(cell::Lattice, expansion::AbstractVector{<:Integer})

Return a supercell based on cell and expansion coefficients.