API Reference

API Reference

Lattice and Cell

Reciprocal space

Metric tensor

Transformations

Others

Lattice and Cell

CrystallographyBase.CrystalSystem — Module

Represent the 7 crystal systems.

CrystallographyBase.LatticeSystem — Module

Represent the 7 lattice systems.

CrystallographyBase.Bravais — Module

Represent the 14 Bravais lattices.

CrystallographyBase.AbstractLattice — Type

Represent the real lattices and the reciprocal lattices.

CrystallographyBase.Lattice — Type

Lattice(mat::AbstractMatrix)

Construct a Lattice from a matrix.

Note

The basis vectors of the matrix are stored as columns.

Lattice(𝐚::AbstractVector, 𝐛::AbstractVector, 𝐜::AbstractVector)

Construct a Lattice from three basis vectors.

Lattice(basisvectors::AbstractVector{<:AbstractVector})

Construct a Lattice from a vector of three basis vectors.

Lattice(a, b, c, α, β, γ; axis = :a)

Construct a Lattice from the six cell parameters.

The default convention we used here is that edge vector 𝐚 in the positive x-axis direction, edge vector 𝐛 in the x-y plane with a positive y-axis component, and edge vector 𝐜 with a positive z-axis component in the Cartesian system. See Wikipedia. You can also choose axis = :c.

Lattice(cell::Cell)

Get the lattice of a Cell.

Lattice(g::MetricTensor)

Construct a Lattice from a MetricTensor.

CrystallographyBase.isrighthanded — Function

isrighthanded(lattice::Lattice)

Test whether the basis vectors are defined to be right-handed.

The basis vectors are right-handed if and only if

\[\mathbf{a} \cdot (\mathbf{b} \times \mathbf{c}) > 0.\]

CrystallographyBase.basisvectors — Function

basisvectors(lattice::Lattice)

Get the three primitive vectors from a lattice.

basisvectors(lattice::ReciprocalLattice)

Get the three basis vectors from a ReciprocalLattice.

CrystallographyBase.latticesystem — Function

latticesystem(bravais::Bravais)

Get the lattice system of a Bravais lattice.

latticesystem(lattice::Lattice; angletol=1e-5, lengthtol=1e-5)

Get the lattice system of a Lattice.

CrystallographyBase.latticeconstants — Function

latticeconstants(lattice::Lattice)

Get the six lattice constants from a lattice.

latticeconstants(g::MetricTensor)

Get the six lattice constants from a MetricTensor.

CrystallographyBase.periodicity — Function

periodicity(lattice::Lattice)

Get crystal periodicity in $x$, $y$, and $z$ direction from the Lattice.

CrystallographyBase.Cell — Type

Cell(lattice, positions, atoms)

Create a new cell.

Argument lattice is a Lattice type. Fractional atomic positions positions are given by a vector of $N$ vectors with floating point values, where $N$ is the number of atoms. Argument atoms is a list of $N$ values, where the same kind of atoms need to be the same type.

CrystallographyBase.supercell — Function

supercell(lattice::Lattice, repfactors::AbstractMatrix{<:Integer})
supercell(lattice::Lattice, repfactors::AbstractVector{<:Integer})
supercell(lattice::Lattice, repfactor::Integer)

Create a supercell from lattice.

supercell(cell::Cell, repfactors::AbstractMatrix{<:Integer})
supercell(cell::Cell, repfactors::AbstractVector{<:Integer})
supercell(cell::Cell, repfactor::Integer)

Create a supercell from cell.

Note

Currently, only integral replications are supported.

CrystallographyBase.eachatom — Function

eachatom(cell::Cell)

Create a generator that iterates over the atoms in a Cell.

Reciprocal space

Note that we take $2\pi$ as $1$, not the solid-state physics convention.

CrystallographyBase.ReciprocalPoint — Type

ReciprocalPoint(x, y, z, w)

Represent a special point of the 3D Brillouin zone. Each of them has a weight w.

CrystallographyBase.ReciprocalLattice — Type

ReciprocalLattice(mat::SMatrix)

Construct a ReciprocalLattice.

Warning

You should not use this function directly, always use reciprocal of a Lattice.

CrystallographyBase.MonkhorstPackGrid — Type

MonkhorstPackGrid(mesh, is_shift)

Represent the Monkhorst–Pack grid.

Arguments

mesh: A length-three vector specifying the k-point grid ($nk_1 × nk_2 × nk_3$) as in Monkhorst–Pack grids.

is_shift: A length-three vector specifying whether the grid is displaced by half a grid step in the corresponding directions.

CrystallographyBase.reciprocal — Function

reciprocal(lattice::Lattice)
reciprocal(lattice::ReciprocalLattice)

Get the reciprocal of a Lattice or a ReciprocalLattice.

CrystallographyBase.coordinates — Function

coordinates(arr::AbstractArray{<:ReciprocalPoint})

Get the coordinates of an array of ReciprocalPoints.

CrystallographyBase.weights — Function

weights(arr::AbstractArray{<:ReciprocalPoint})

Get the weights of an array of ReciprocalPoints.

Metric tensor

CrystallographyBase.MetricTensor — Type

MetricTensor(𝐚::AbstractVector, 𝐛::AbstractVector, 𝐜::AbstractVector)

Generate a MetricTensor from the three basis vectors.

MetricTensor(lattice::Lattice)

Generate a MetricTensor from a Lattice.

MetricTensor(a, b, c, α, β, γ)

Generate a MetricTensor from the six cell parameters.

CrystallographyBase.distance — Function

distance(𝐚::AbstractVector, g::MetricTensor, 𝐛::AbstractVector)

Get the distance between two coordinates using a MetricTensor.

Transformations

CrystallographyBase.CartesianFromFractional — Type

CartesianFromFractional(lattice::Union{Lattice,ReciprocalLattice})
CartesianFromFractional(a, b, c, α, β, γ)

Get the transformation from fractional coordinates to Cartesian coordinates.

CrystallographyBase.FractionalFromCartesian — Type

FractionalFromCartesian(lattice::Union{Lattice,ReciprocalLattice})
FractionalFromCartesian(a, b, c, α, β, γ)

Get the transformation from Cartesian coordinates to fractional coordinates.

CrystallographyBase.PrimitiveFromStandardized — Type

PrimitiveFromStandardized(tf::AbstractMatrix)

Construct the transformation from a standardized cell to a primitive cell.

CrystallographyBase.StandardizedFromPrimitive — Type

StandardizedFromPrimitive(tf::AbstractMatrix)

Construct the transformation from a primitive cell to a standardized cell.

Others

CrystallographyBase.cellvolume — Function

cellvolume(a, b, c, α, β, γ)

Calculate the cell volume from 6 cell parameters.

cellvolume(l::Lattice)
cellvolume(c::Cell)

Calculate the cell volume from a Lattice or a Cell.

cellvolume(g::MetricTensor)

Calculate the cell volume from a MetricTensor.

CrystallographyBase.crystaldensity — Function

crystaldensity(lattice::Lattice, atoms)
crystaldensity(cell::Cell)

Calculate the density of a crystal structure.

Here, atoms is an iterable of atomic numbers, element names, symbols, or Mendeleev.Elements. You can extend the atomicmass method to work with custom types.