API

API

Lattice

Reciprocal space

Miller and Miller–Bravais indices

Metric tensor

Transformations

Others

Lattice

CrystallographyBase.CrystalSystem — Type

Represent one of the seven crystal systems.

CrystallographyBase.Triclinic — Type

Triclinic()

Represent the triclinic system.

CrystallographyBase.Monoclinic — Type

Monoclinic()

Represent the monoclinic system.

CrystallographyBase.Orthorhombic — Type

Orthorhombic()

Represent the orthorhombic system.

CrystallographyBase.Tetragonal — Type

Tetragonal()

Represent the tetragonal system.

CrystallographyBase.Cubic — Type

Cubic()

Represent the cubic system.

CrystallographyBase.Trigonal — Type

Trigonal()

Represent the trigonal system.

CrystallographyBase.Hexagonal — Type

Hexagonal()

Represent the hexagonal system.

CrystallographyBase.Centering — Type

Represent the centering types.

CrystallographyBase.BaseCentering — Type

BaseCentering{:A}()
BaseCentering{:B}()
BaseCentering{:C}()

Represent the base-centering.

CrystallographyBase.Primitive — Type

Primitive()

Represent no centering.

CrystallographyBase.BodyCentering — Type

BodyCentering()

Represent the body-centering.

CrystallographyBase.FaceCentering — Type

FaceCentering()

Represent the face-centering.

CrystallographyBase.RhombohedralCentering — Type

RhombohedralCentering()

Represent the rhombohedral-centering of the hexagonal system.

Missing docstring.

Missing docstring for BaseCentering. Check Documenter's build log for details.

CrystallographyBase.Bravais — Type

Bravais(a::CrystalSystem, b::Centering, obverse::Bool=true)

Represent a Bravais lattice type.

CrystallographyBase.Lattice — Type

Lattice(mat::AbstractMatrix)

Construct a Lattice from a matrix.

Note

The basis vectors of the matrix are stored as columns.

Lattice(𝐚::AbstractVector, 𝐛::AbstractVector, 𝐜::AbstractVector)

Construct a Lattice from three basis vectors.

Lattice(cell::Cell)

Get the lattice of a Cell.

Lattice(a, b, c, α, β, γ)

Construct a Lattice from the six cell parameters.

The convention we used here is that edge vector 𝐚 in the positive x-axis direction, edge vector 𝐛 in the x-y plane with positive y-axis component, and edge vector 𝐜 with positive z-axis component in the Cartesian-system. See Wikipedia.

Lattice(g::MetricTensor)

Construct a Lattice from a MetricTensor.

CrystallographyBase.centering — Function

centering(bravais::Bravais)

Get the centering type of a Bravais type.

CrystallographyBase.crystalsystem — Function

crystalsystem(bravais::Bravais)

Get the crystal system of a Bravais type.

crystalsystem(a, b, c, α, β, γ)

Guess the crystal system from the six cell parameters.

crystalsystem(lattice::Lattice)

Get the crystal system of a lattice.

Spglib.basis_vectors — Function

basis_vectors(lattice::Lattice)

Get the three basis vectors from a lattice.

basis_vectors(lattice::ReciprocalLattice)

Get the three basis vectors from a ReciprocalLattice.

CrystallographyBase.cellparameters — Function

cellparameters(lattice::Lattice)

Get the six cell parameters from a lattice.

cellparameters(g::MetricTensor)

Get the six cell parameters from a MetricTensor.

CrystallographyBase.supercell — Function

supercell(lattice::Lattice, expansion::AbstractMatrix{<:Integer})

Allow the supercell to be a tilted extension of cell.

supercell(lattice::Lattice, expansion::AbstractVector{<:Integer})

Return a supercell based on cell and expansion coefficients.

Reciprocal space

Note that we take $2\pi$ as $1$, not the solid-state physics convention.

CrystallographyBase.ReciprocalPoint — Type

ReciprocalPoint(x, y, z, w)

Represent a special point of the 3D Brillouin zone. Each of them has a weight w.

CrystallographyBase.ReciprocalLattice — Type

ReciprocalLattice(mat::SMatrix)

Construct a ReciprocalLattice.

Warning

You should not use this function directly, always use inv of a Lattice.

CrystallographyBase.MonkhorstPackGrid — Type

MonkhorstPackGrid(mesh, is_shift)

Represent the Monkhorst–Pack grid.

Arguments

mesh: A length-three vector specifying the k-point grid ($nk_1 × nk_2 × nk_3$) as in Monkhorst–Pack grids.

is_shift: A length-three vector specifying whether the grid is displaced by half a grid step in the corresponding directions.

Base.inv — Function

inv(lattice::Lattice)
inv(lattice::ReciprocalLattice)

Get the reciprocal of a Lattice or a ReciprocalLattice.

CrystallographyBase.reciprocal_mesh — Function

reciprocal_mesh(cell::Cell, mesh, is_shift; kwargs...)

List the ReciprocalPoints from the mesh of the reciprocal space of a Cell.

Arguments

cell::Cell: the cell.

mesh: a vector of three integers which specify the mesh numbers along reciprocal primitive axis.

is_shift=falses(3): a vector of three elements specifying whether the mesh is shifted along the axis in half of adjacent mesh points irrespective of the mesh numbers. The allowed values are 0, 1, true, and false.

is_time_reversal=true: whether to impose the time reversal symmetry on the mesh.

symprec=1e-5: distance tolerance in Cartesian coordinates to find crystal symmetry.

cartesian=false: whether to return the reciprocal points in Cartesian coordinates.

ir_only=true: whether to return the reciprocal points only in the irreducible Brillouin zone.

CrystallographyBase.coordinates — Function

coordinates(arr::AbstractArray{<:ReciprocalPoint})

Get the coordinates of an array of ReciprocalPoints.

CrystallographyBase.weights — Function

weights(arr::AbstractArray{<:ReciprocalPoint})

Get the weights of an array of ReciprocalPoints.

Missing docstring.

Missing docstring for vertices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for faces. Check Documenter's build log for details.

Miller and Miller–Bravais indices

CrystallographyBase.Miller — Type

Miller(i, j, k)

Represent the Miller indices in the real space (crystal directions).

CrystallographyBase.MillerBravais — Type

MillerBravais(i, j, k, l)

Represent the Miller–Bravais indices in the real space (crystal directions).

CrystallographyBase.ReciprocalMiller — Type

ReciprocalMiller(i, j, k)

Represent the Miller indices in the reciprocal space (planes).

CrystallographyBase.ReciprocalMillerBravais — Type

ReciprocalMillerBravais(i, j, k, l)

Represent the Miller–Bravais indices in the reciprocal space (planes).

CrystallographyBase.family — Function

family(x::Union{Miller,MillerBravais,ReciprocalMiller,ReciprocalMillerBravais})

List the all the directions/planes that are equivalent to x by symmetry.

CrystallographyBase.@m_str — Macro

m_str(s)

Generate the Miller indices or Miller–Bravais indices quickly.

Examples

julia> m"[-1, 0, 1]"
3-element Miller:
 -1
  0
  1

julia> m"<2, -1, -1, 3>"
4-element MillerBravais:
  2
 -1
 -1
  3

julia> m"(-1, 0, 1)"
3-element ReciprocalMiller:
 -1
  0
  1

julia> m"(1, 0, -1, 0)"
4-element ReciprocalMillerBravais:
  1
  0
 -1
  0

Metric tensor

CrystallographyBase.MetricTensor — Type

MetricTensor(𝐚::AbstractVector, 𝐛::AbstractVector, 𝐜::AbstractVector)

Generate a MetricTensor from the three basis vectors.

MetricTensor(lattice::Lattice)

Generate a MetricTensor from a Lattice.

MetricTensor(a, b, c, α, β, γ)

Generate a MetricTensor from the six cell parameters.

CrystallographyBase.directioncosine — Function

directioncosine(𝐚::AbstractVector, g::MetricTensor, 𝐛::AbstractVector)

Get the direction cosine of two vectors and a MetricTensor.

CrystallographyBase.directionangle — Function

directionangle(𝐚::AbstractVector, g::MetricTensor, 𝐛::AbstractVector)

Get the direction angle of two vectors and a MetricTensor.

CrystallographyBase.distance — Function

distance(𝐚::AbstractVector, g::MetricTensor, 𝐛::AbstractVector)

Get the distance between two coordinates using a MetricTensor.

CrystallographyBase.interplanar_spacing — Function

interplanar_spacing(𝐚::AbstractVector, g::MetricTensor)

Get the interplanar spacing from a MetricTensor.

Transformations

CrystallographyBase.CartesianFromFractional — Type

CartesianFromFractional(lattice::Union{Lattice,ReciprocalLattice})
CartesianFromFractional(a, b, c, α, β, γ)

Get the transformation from fractional coordinates to Cartesian coordinates.

CrystallographyBase.FractionalFromCartesian — Type

FractionalFromCartesian(lattice::Union{Lattice,ReciprocalLattice})
FractionalFromCartesian(a, b, c, α, β, γ)

Get the transformation from Cartesian coordinates to fractional coordinates.

CrystallographyBase.PrimitiveFromStandardized — Type

PrimitiveFromStandardized(tf::AbstractMatrix)

Construct the transformation from a standardized cell to a primitive cell.

CrystallographyBase.StandardizedFromPrimitive — Type

StandardizedFromPrimitive(tf::AbstractMatrix)

Construct the transformation from a primitive cell to a standardized cell.

Others

CrystallographyBase.cellvolume — Function

cellvolume(a, b, c, α, β, γ)

Calculates the cell volume from 6 cell parameters.

cellvolume(l::Lattice)
cellvolume(c::Cell)

Calculates the cell volume from a Lattice or a Cell.

cellvolume(g::MetricTensor)

Calculates the cell volume from a MetricTensor.